used authorize helper instead of gate

+++ app/Http/Controllers/PostsController.php
@@ -75,9 +75,10 @@ class PostsController extends Controller
     {
         $post = BlogPost::findOrFail($id);
 
-        if (Gate::denies('update-post', $post)) {
-            abort(403, "You can't edit this blog post.");
-        }
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post.");
+        // }
+        $this->authorize('update-post', $post);
 
         return view('posts.edit', ['post' => $post]);
     }
@@ -93,9 +94,10 @@ class PostsController extends Controller
     {
         $post = BlogPost::findOrFail($id);
 
-        if (Gate::denies('update-post', $post)) {
-            abort(403, "You can't edit this blog post.");
-        }
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post.");
+        // }
+        $this->authorize('update-post', $post);
 
         $validated = $request->validated();
         $post->fill($validated);
@@ -114,6 +116,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post.");
+        // }
+        $this->authorize('delete-post', $post);
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
